naming, imports, and argparsers for one cli-entrypoint
