+++ .gitignore
@@ -1,6 +1,5 @@
 *.dbm
-*/data/*
-!*/data/*.xlsx
+*/data-folder/*
 
 # Created by https://www.gitignore.io
