+++ tests/test_graph.py
@@ -1,5 +1,5 @@
 import unittest
-from wikimap import wikimap
+from wikimap import graph
 import networkx.testing as nxt
 import networkx as nx
 
@@ -7,7 +7,7 @@ import networkx as nx
 class TestGeneralNetworkMethods(unittest.TestCase):
 
     def setUp(self):
-        self.G = wikimap.WikiMap()
+        self.G = graph.WikiMap()
 
         # four node group
         self.G.add_edge("A", "B")
@@ -35,17 +35,17 @@ class TestGeneralNetworkMethods(unittest.TestCase):
 
     def test_connected_components_with_size(self):
         # three node group
-        expected_three1 = wikimap.WikiMap()
+        expected_three1 = graph.WikiMap()
         expected_three1.add_edge("Y", "X")
         expected_three1.add_edge("Z", "X")
 
         # three node group (another)
-        expected_three2 = wikimap.WikiMap()
+        expected_three2 = graph.WikiMap()
         expected_three2.add_edge("alpha", "beta")
         expected_three2.add_edge("alpha", "gamma")
 
         # four node group
-        expected_four = wikimap.WikiMap()
+        expected_four = graph.WikiMap()
         expected_four.add_edge("A", "B")
         expected_four.add_edge("B", "A")
         expected_four.add_edge("A", "C")
@@ -68,55 +68,55 @@ class TestGeneralNetworkMethods(unittest.TestCase):
 class TestCleaningNodes(unittest.TestCase):
 
     def test_clean_skips(self):
-        self.assertEqual(wikimap.WikiMap.clean("File: foo"), "File: foo")
+        self.assertEqual(graph.WikiMap.clean("File: foo"), "File: foo")
         self.assertEqual(
-            wikimap.WikiMap.clean("!!!!!foo!!!!!"), "!!!!!foo!!!!!")
+            graph.WikiMap.clean("!!!!!foo!!!!!"), "!!!!!foo!!!!!")
 
     def test_clean_unicode(self):
         self.assertEqual(
-            wikimap.WikiMap.clean(u'Rate\xa0of\xa0fire'), "rate of fire")
+            graph.WikiMap.clean(u'Rate\xa0of\xa0fire'), "rate of fire")
 
     def test_clean_punct_remove(self):
         # capitalization and punctuation removal
-        self.assertEqual(wikimap.WikiMap.clean("Heights"), "heights")
+        self.assertEqual(graph.WikiMap.clean("Heights"), "heights")
         self.assertEqual(
-            wikimap.WikiMap.clean("discovery_site"), "discovery site")
-        self.assertEqual(wikimap.WikiMap.clean("Max. devices"), "max devices")
-        self.assertEqual(wikimap.WikiMap.clean("Circus tent?"), "circus tent")
-        self.assertEqual(wikimap.WikiMap.clean("Web site:"), "web site")
-        self.assertEqual(wikimap.WikiMap.clean("/Karaoke"), "karaoke")
+            graph.WikiMap.clean("discovery_site"), "discovery site")
+        self.assertEqual(graph.WikiMap.clean("Max. devices"), "max devices")
+        self.assertEqual(graph.WikiMap.clean("Circus tent?"), "circus tent")
+        self.assertEqual(graph.WikiMap.clean("Web site:"), "web site")
+        self.assertEqual(graph.WikiMap.clean("/Karaoke"), "karaoke")
 
     def test_clean_punct_except(self):
         # exception to punctuation removal
-        self.assertEqual(wikimap.WikiMap.clean("% of total exports"),
+        self.assertEqual(graph.WikiMap.clean("% of total exports"),
                          "% of total exports")
-        self.assertEqual(wikimap.WikiMap.clean("Managing editor, design"),
+        self.assertEqual(graph.WikiMap.clean("Managing editor, design"),
                          "managing editor, design")
-        self.assertEqual(wikimap.WikiMap.clean("MSRP US$"), "msrp us$")
-        self.assertEqual(wikimap.WikiMap.clean("Specific traits & abilities"),
+        self.assertEqual(graph.WikiMap.clean("MSRP US$"), "msrp us$")
+        self.assertEqual(graph.WikiMap.clean("Specific traits & abilities"),
                          "specific traits & abilities")
-        self.assertEqual(wikimap.WikiMap.clean("re-issuing"), "re-issuing")
+        self.assertEqual(graph.WikiMap.clean("re-issuing"), "re-issuing")
         # another good example: "Capital-in-exile"
 
     def test_clean_HTML_remove(self):
         # HTML-like junk removal
-        self.assertEqual(wikimap.WikiMap.clean("<hiero>G16</hiero>"), "g16")
-        self.assertEqual(wikimap.WikiMap.clean("&mdot;foo"), "foo")
-        self.assertEqual(wikimap.WikiMap.clean("Opened</th>"), "opened")
+        self.assertEqual(graph.WikiMap.clean("<hiero>G16</hiero>"), "g16")
+        self.assertEqual(graph.WikiMap.clean("&mdot;foo"), "foo")
+        self.assertEqual(graph.WikiMap.clean("Opened</th>"), "opened")
 
     def test_clean_parens(self):
         self.assertEqual(
-            wikimap.WikiMap.clean("Parent club(s)"), "parent club")
+            graph.WikiMap.clean("Parent club(s)"), "parent club")
         self.assertEqual(
-            wikimap.WikiMap.clean("Team president (men)"), "team president")
-        self.assertEqual(wikimap.WikiMap.clean("Deaconess(es)"), "deaconess")
-        self.assertEqual(wikimap.WikiMap.clean("ARWU[5]"), "arwu")
+            graph.WikiMap.clean("Team president (men)"), "team president")
+        self.assertEqual(graph.WikiMap.clean("Deaconess(es)"), "deaconess")
+        self.assertEqual(graph.WikiMap.clean("ARWU[5]"), "arwu")
 
     def test_clean_possessive(self):
         self.assertEqual(
-            wikimap.WikiMap.clean("Women's coach"), "women's coach")
+            graph.WikiMap.clean("Women's coach"), "women's coach")
         self.assertEqual(
-            wikimap.WikiMap.clean("Teams' champion"), "teams' champion")
+            graph.WikiMap.clean("Teams' champion"), "teams' champion")
 
 
 class TestAddToField(unittest.TestCase):
@@ -125,21 +125,21 @@ class TestAddToField(unittest.TestCase):
         self.test_dict = ["bannana", {"foo": ["bar"]}]
 
     def test_add_to_existing_field_new(self):
-        wikimap.WikiMap.add_to_field(self.test_dict[1], "foo", "aba")
+        graph.WikiMap.add_to_field(self.test_dict[1], "foo", "aba")
         self.assertItemsEqual(self.test_dict[1], {"foo": ["bar", "aba"]})
 
     def test_add_to_existing_field_existing(self):
-        wikimap.WikiMap.add_to_field(self.test_dict[1], "foo", "bar")
+        graph.WikiMap.add_to_field(self.test_dict[1], "foo", "bar")
         self.assertItemsEqual(self.test_dict[1], {"foo": ["bar"]})
 
     def test_add_to_new_field(self):
-        wikimap.WikiMap.add_to_field(self.test_dict[1], "bar", "aba")
+        graph.WikiMap.add_to_field(self.test_dict[1], "bar", "aba")
         self.assertItemsEqual(
             self.test_dict[1], {"foo": ["bar"], "bar": ["aba"]})
 
     def test_add_to_new_location(self):
         self.test_dict[1]['infobox'] = {}
-        wikimap.WikiMap.add_to_field(
+        graph.WikiMap.add_to_field(
             self.test_dict[1]['infobox'], "bar", "aba")
         self.assertItemsEqual(self.test_dict[1], {"foo": ["bar"],
                                                   "infobox": {"bar": ["aba"]}})
@@ -148,7 +148,7 @@ class TestAddToField(unittest.TestCase):
 class TestInsertingInformation(unittest.TestCase):
 
     def setUp(self):
-        self.G = wikimap.WikiMap()
+        self.G = graph.WikiMap()
 
     def test_add_uncleaned(self):
         self.G.add_uncleaned('_unrend_', '_rend_')
@@ -237,7 +237,7 @@ class TestInsertingInformation(unittest.TestCase):
 class TestFetchingInformation(unittest.TestCase):
 
     def setUp(self):
-        self.G = wikimap.WikiMap()
+        self.G = graph.WikiMap()
         self.G.add_edge('unrend', 'hybrid')
         self.G.add_edge('hybrid', 'rend')
